+++ internal/server/authentication.go
@@ -1,0 +1,76 @@
+package server
+
+import (
+	"net/http"
+	"strings"
+
+	"github.com/gin-contrib/sessions"
+	"github.com/gin-contrib/sessions/cookie"
+	"github.com/gin-gonic/gin"
+)
+
+var secret = []byte("secret")
+
+const userKey = "user"
+
+func (s *Server) RegisterAuthenticationRoutes(r *gin.Engine) *gin.Engine {
+	r.Use(sessions.Sessions("mysession", cookie.NewStore([]byte(s.env.Secret))))
+
+	r.POST("/login", s.Login)
+	r.GET("/logout", s.Logout)
+
+	return r
+}
+
+func (s *Server) AuthRequired(c *gin.Context) {
+	session := sessions.Default(c)
+	user := session.Get(userKey)
+
+	if user == nil {
+		c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
+		return
+	}
+
+	c.Next()
+}
+
+func (s *Server) Login(c *gin.Context) {
+	session := sessions.Default(c)
+	username := c.PostForm("username")
+	password := c.PostForm("password")
+
+	if strings.Trim(username, " ") == "" || strings.Trim(password, " ") == "" {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Parameters can't be empty"})
+		return
+	}
+
+	if username != "hello" || password != "world" {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "Authentication failed"})
+		return
+	}
+
+	session.Set(userKey, username)
+	if err := session.Save(); err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to save session"})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "Successfully authenticated user"})
+}
+
+func (s *Server) Logout(c *gin.Context) {
+	session := sessions.Default(c)
+	user := session.Get(userKey)
+	if user == nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid session token"})
+		return
+	}
+
+	session.Delete(userKey)
+	if err := session.Save(); err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to save session"})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"message": "Successfully logged out"})
+}
