+++ internal/server/routes.go
@@ -2,19 +2,11 @@ package server
 
 import (
 	"net/http"
-	"strings"
 
 	"github.com/gin-contrib/cors"
-	"github.com/gin-contrib/sessions"
-	"github.com/gin-contrib/sessions/cookie"
 	"github.com/gin-gonic/gin"
 	"github.com/slugger7/exorcist/internal/environment"
-	"github.com/slugger7/exorcist/internal/job"
 )
-
-var secret = []byte("secret")
-
-const userKey = "user"
 
 func (s *Server) RegisterRoutes() http.Handler {
 	if s.env.AppEnv == environment.AppEnvEnum.Prod {
@@ -31,82 +23,19 @@ func (s *Server) RegisterRoutes() http.Handler {
 		AllowCredentials: true, // Enable cookies/auth
 	}))
 
-	r.Use(sessions.Sessions("mysession", cookie.NewStore(secret)))
+	r = s.RegisterAuthenticationRoutes(r)
 
 	authenticated := r.Group("/api")
-	authenticated.Use(AuthRequired)
-	{
-		authenticated.GET("/users", s.getUsers)
-	}
+	authenticated.Use(s.AuthRequired)
+	authenticated = s.RegisterUserRoutes(authenticated)
 
-	r.POST("/login", s.login)
-	r.GET("/logout", s.logout)
+	r.GET("/health", s.HealthHandler)
 	r.GET("/", s.HelloWorldHandler)
-
-	r.GET("/health", s.healthHandler)
-
-	r.GET("/library/scan", s.scanLibrary)
-
 	return r
 }
 
-func AuthRequired(c *gin.Context) {
-	session := sessions.Default(c)
-	user := session.Get(userKey)
-
-	if user == nil {
-		c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
-		return
-	}
-
-	c.Next()
-}
-
-func (s *Server) login(c *gin.Context) {
-	session := sessions.Default(c)
-	username := c.PostForm("username")
-	password := c.PostForm("password")
-
-	if strings.Trim(username, " ") == "" || strings.Trim(password, " ") == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Parameters can't be empty"})
-		return
-	}
-
-	if username != "hello" || password != "world" {
-		c.JSON(http.StatusUnauthorized, gin.H{"error": "Authentication failed"})
-		return
-	}
-
-	session.Set(userKey, username)
-	if err := session.Save(); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to save session"})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{"message": "Successfully authenticated user"})
-}
-
-func (s *Server) logout(c *gin.Context) {
-	session := sessions.Default(c)
-	user := session.Get(userKey)
-	if user == nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid session token"})
-		return
-	}
-
-	session.Delete(userKey)
-	if err := session.Save(); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to save session"})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{"message": "Successfully logged out"})
-}
-
-func (s *Server) getUsers(c *gin.Context) {
-	session := sessions.Default(c)
-	user := session.Get(userKey)
-	c.JSON(http.StatusOK, gin.H{"user": user})
+func (s *Server) HealthHandler(c *gin.Context) {
+	c.JSON(http.StatusOK, s.db.Health())
 }
 
 func (s *Server) HelloWorldHandler(c *gin.Context) {
@@ -114,12 +43,4 @@ func (s *Server) HelloWorldHandler(c *gin.Context) {
 	resp["message"] = "Hello World"
 
 	c.JSON(http.StatusOK, resp)
-}
-
-func (s *Server) healthHandler(c *gin.Context) {
-	c.JSON(http.StatusOK, s.db.Health())
-}
-
-func (s *Server) scanLibrary(c *gin.Context) {
-	go job.ScanPath(s.db)
 }
