+++ internal/server/users.go
@@ -1,0 +1,20 @@
+package server
+
+import (
+	"net/http"
+
+	"github.com/gin-contrib/sessions"
+	"github.com/gin-gonic/gin"
+)
+
+func (s *Server) RegisterUserRoutes(r *gin.RouterGroup) *gin.RouterGroup {
+	r.GET("/users", s.GetUsers)
+
+	return r
+}
+
+func (s *Server) GetUsers(c *gin.Context) {
+	session := sessions.Default(c)
+	user := session.Get(userKey)
+	c.JSON(http.StatusOK, gin.H{"user": user})
+}
