+++ template.env
@@ -6,5 +6,6 @@ DATABASE_NAME=exorcist
 DATABASE_CONNECTION_STRING=postgres://${DATABASE_USER}:${DATABASE_PASSWORD}@${DATABASE_HOST}:${DATABASE_PORT}?sslmode=disable
 MEDIA_PATH=./
 DEBUG_SQL=true
-APP_ENV=local
+APP_ENV=production
 PORT=8080
+SECRET=some-secure-secret
